Fixed 2 splits
Chapter 07 - Changed location just a bit so it isnt as awkward on the
landing on the tree.

Chapter 28 - Changed y location so it splits even when you are Parhelion
(or are too fast in general).

Misc - version number doesnt disable the script & added a space in
chapter 26 for visual clarity.